model extensions wip
