+++ src/SFA.DAS.EAS.Portal.Worker/EventHandlers/Commitments/CohortApprovalRequestedByProviderEventHandler.cs
@@ -43,7 +43,7 @@ namespace SFA.DAS.EAS.Portal.Worker.EventHandlers.Commitments
                 organisation.Name = commitment.LegalEntityName;
             }
 
-            var (cohort, cohortCreated) = GetOrAddCohort(organisation, cohortApprovalRequestedByProvider.CommitmentId);
+            var (cohort, cohortCreated) = organisation.GetOrAddCohort(cohortApprovalRequestedByProvider.CommitmentId);
             if (cohortCreated == EntityCreation.Created)
             {
                 cohort.Reference = commitment.Reference;
